+++ src/App.tsx
@@ -2,7 +2,8 @@ import Wrapper from "./components/UI/Wrapper";
 import Puzzle from "./components/Puzzle";
 import Header from "./components/Header";
 import { useReducer, Reducer, useEffect } from "react";
-import puzzleReducer, { initialState as initialAttempts, State, Action, colCount, rowCount } from "./store/puzzleReducer";
+import puzzleReducer, { initialState as initialAttempts, State, colCount, rowCount } from "./store/puzzleReducer";
+import { Action } from "./store/action";
 function App() {
   const answer:string[]=["G", "R", "A", "I", "L"];
   const [state, dispatch] = useReducer<Reducer<State, Action>>(puzzleReducer, initialAttempts);
